requirements.txt
ansible==1.9.1
astroid==1.3.6
behave==1.2.5
colorama==0.3.3
coverage==3.7.1
Django==1.8.1
django-angular-scaffold==0.12.1
django-appconf==1.0.1
django-behave==0.1.4
django-compressor==1.5
django-extensions==1.5.5
django-rest-swagger==0.2.9
django-reversion==1.8.6
django-uuidfield==0.5.0
djangorestframework==3.1.2
dodgy==0.1.7
drf-extensions==0.2.7
ecdsa==0.13
enum34==1.0.4
Fabric==1.10.1
gitdb==0.6.4
GitPython==1.0.1
Jinja2==2.7.3
logilab-common==0.63.2
mando==0.3.2
MarkupSafe==0.23
mccabe==0.3
MySQL-python==1.2.5
paramiko==1.15.2
parse==1.6.6
parse-type==0.3.4
pep257==0.5.0
pep8==1.6.2
pep8-naming==0.2.2
prospector==0.10.2
pycrypto==2.6.1
pyflakes==0.8.1
pylint==1.4.3
pylint-celery==0.3
pylint-common==0.2.1
pylint-django==0.6
pylint-plugin-utils==0.2.3
python-coveralls==2.5.0
PyYAML==3.11
radon==1.0
requests==2.7.0
requirements-detector==0.4
selenium==2.45.0
setoptconf==0.2.0
sh==1.11
six==1.9.0
smmap==0.9.0
splinter==0.7.2
xenon==0.3.1
